calc-cbs-params: Fix division operation
In Python 3, the operator '/' always return a float value so the
calc_sr_class_params() function returns 'idleslope' as float type, not
as integer as expected. To fix it, this patch uses the '//' operator,
instead.

As a side note, the division in 'idleslope' calculation is expected to
be an exact division always since 'stream[rate]' is a multiple of 4000,
so no precision error should be introduced by this patch.